# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  models.InfobaseBlocker:
    properties:
      cluster_id:
        example: 80f7f2f6-2feb-46bf-92f4-19294a2f5dc7
        type: string
      denied_from:
        example: "2020-10-01T08:30:00Z"
        type: string
      denied_parameter:
        example: code
        type: string
      denied_to:
        example: "2020-10-01T08:30:00Z"
        type: string
      infobase:
        example: testdb2
        type: string
      message:
        example: Обновление информационной базы
        type: string
      permission_code:
        example: "123"
        type: string
      reload:
        type: boolean
      scheduled_jobs_deny:
        type: boolean
      sessions_deny:
        type: boolean
    type: object
  models.InfobaseUnblocker:
    properties:
      cluster_id:
        example: 80f7f2f6-2feb-46bf-92f4-19294a2f5dc7
        type: string
      denied_parameter:
        type: string
      infobase:
        example: testdb2
        type: string
      infobase_id:
        example: 80f7f2f6-2feb-46bf-92f4-19294a2f5dc7
        type: string
      permission_code:
        type: string
      scheduled_jobs_deny:
        type: boolean
      sessions_deny:
        type: boolean
    type: object
...
  serialize.InfobaseSummaryInfo:
    properties:
      cluster_id:
        type: string
      descr:
        example: УППБоеваяБаза
        type: string
      name:
        example: УППБоеваяБаза
        type: string
      uuid:
        example: efa3672f-947a-4d84-bd58-b21997b83561
        type: string
    type: object
...
host: localhost:3001
info:
  contact:
    email: khorevaa@yandex.ru
    name: API Support
  description: KUBOdin Swagger UI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'KUBOdin: Remote Administration for 1S.Enterprise Application Servers'
  version: "1.0"
paths:
...
  /app/{app}/block:
    get:
      consumes:
      - application/json
      description: Установка блокировки на произвольную информационную базу на сервер 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: infobase uuid or name
        in: query
        name: infobase-id
        required: true
        type: string
      - description: session deny
        in: query
        name: sessions-deny
        required: true
        type: string
      - description: message to user
        in: query
        name: message
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: denied parameter
        in: query
        name: denied-parameter
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: scheduled jobs deny
        in: query
        name: scheduled-jobs-deny
        type: boolean
      - default: '"now"'
        description: denied from time
        in: query
        name: denied-from
        type: string
      - description: denied to time
        in: query
        name: denied-to
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на произвольную информационную базу на сервер 1С Предприятие
      tags:
      - blocker
    post:
      consumes:
      - application/json
      description: Установка блокировки на произвольную информационную базу на сервер 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: block info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InfobaseBlocker'
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на произвольную информационную базу на сервер 1С Предприятие
      tags:
      - blocker
...
  /app/{app}/clusters/{cluster}/block:
    get:
      consumes:
      - application/json
      description: установка блокировки на информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase uuid or name
        in: query
        name: infobase
        required: true
        type: string
      - description: session deny
        in: query
        name: sessions-deny
        required: true
        type: string
      - description: message to user
        in: query
        name: message
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: denied parameter
        in: query
        name: denied-parameter
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: scheduled jobs deny
        in: query
        name: scheduled-jobs-deny
        type: boolean
      - description: denied from time
        in: query
        name: denied-from
        type: string
      - description: denied to time
        in: query
        name: denied-to
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на информационную базу на кластере
      tags:
      - blocker
    post:
      consumes:
      - application/json
      description: установка блокировки на информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: block info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InfobaseBlocker'
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на информационную базу на кластере
      tags:
      - blocker
...
  /app/{app}/clusters/{cluster}/infobases/{infobase}/block:
    get:
      consumes:
      - application/json
      description: Установка блокировки на выбранную информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase uuid or name
        in: path
        name: infobase
        required: true
        type: string
      - description: session deny
        in: query
        name: sessions-deny
        required: true
        type: string
      - description: message to user
        in: query
        name: message
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: denied parameter
        in: query
        name: denied-parameter
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: scheduled jobs deny
        in: query
        name: scheduled-jobs-deny
        type: boolean
      - description: denied from time
        in: query
        name: denied-from
        type: string
      - description: denied to time
        in: query
        name: denied-to
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на выбранную информационную базу на кластере
      tags:
      - blocker
    post:
      consumes:
      - application/json
      description: установка блокировки на конкретную информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase uuid or name
        in: path
        name: infobase
        required: true
        type: string
      - description: block info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InfobaseBlocker'
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InfobaseUnblocker'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Установка блокировки на конкретную информационную базу на кластере
      tags:
      - blocker
...
  /app/{app}/clusters/{cluster}/unblock:
    get:
      consumes:
      - application/json
      description: Снятие блокировки на информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase uuid or name
        in: query
        name: infobase
        required: true
        type: string
      - description: session deny
        in: query
        name: sessions-deny
        required: true
        type: string
      - description: denied parameter
        in: query
        name: denied-parameter
        type: string
      - description: permission code
        in: query
        name: permission-code
        type: string
      - description: scheduled jobs deny
        in: query
        name: scheduled-jobs-deny
        type: boolean
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseSummaryInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Снятие блокировки на информационную базу на кластере
      tags:
      - blocker
    post:
      consumes:
      - application/json
      description: Снимает блокировку на информационную базу на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: unblock info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InfobaseUnblocker'
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseSummaryInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: Снимает блокировку на информационную базу на кластере
      tags:
      - blocker
...
swagger: "2.0"